redirect user to homepage if they already have sesison
