+++ src/interfaces/auth.ts
@@ -1,3 +1,8 @@
+export interface FormDataErrorProps {
+  path: string;
+  message: string;
+}
+
 export interface LoginBodyProps {
   username: string;
   password: string;
